`global-less-path` add streamTransform utils
