+++ gulp/task-resolve-css-imports.js
@@ -1,23 +1,13 @@
 const gulp = require('gulp');
-const { Transform } = require('stream');
+const { replace } = require('./utils/streamTransform');
 
 const lessImportTemplate = /^((require|import\s)\(?['"].+)\.less(['"]\)?;)$/gm;
-
-const resolveCss = () =>
-    new Transform({
-        objectMode: true,
-        transform(chunk, _, callback) {
-            const content = chunk.contents.toString();
-            const resolvedContent = content.replace(lessImportTemplate, '$1.css$3');
-            chunk.contents = Buffer.from(resolvedContent);
-            callback(undefined, chunk);
-        },
-    });
+const cssExtensionTemplate = '$1.css$3';
 
 const JS_SRC = ['build/**/*.{js,d.ts}'];
 
 function resolveCssImports() {
-    return gulp.src(JS_SRC).pipe(resolveCss()).pipe(gulp.dest('build'));
+    return gulp.src(JS_SRC).pipe(replace(lessImportTemplate, cssExtensionTemplate)).pipe(gulp.dest('build'));
 }
 
 module.exports = {
