+++ gulp/task-transpile-css.js
@@ -3,18 +3,7 @@ const cached = require('gulp-cached');
 const less = require('gulp-less');
 const postcss = require('gulp-postcss');
 const cssnano = require('cssnano');
-const { Transform } = require('stream');
-
-const filterEmptyFile = () =>
-    new Transform({
-        objectMode: true,
-        transform(file, _, callback) {
-            if (file.contents.length > 0) {
-                this.push(file);
-            }
-            callback();
-        },
-    });
+const { filterEmptyFile } = require('./utils/streamTransform');
 
 const CSS_SRC = ['src/**/*.{css,less}'];
